Move some disciplinary commands to a new file.
Also update the .gitignore file to prevent accidental inclusion of front end files.
